doc: update jquery version to 3.1.2
The jquery provided by Sphinx 1.5.5 has a bug dealing with the
coordinates returned for offset().top for empty <span> tags, generated
by Sphinx for intra-document labels used to create hyperlink anchor
targets.  As reported in PR #5591, some browsers (chrome for example)
don't scroll to the right spot in the document when we link to these
intra-document label points.  jquery v3.1.2 fixes this.  (When Sphinx
updates the jquery included in its release, we can remove our copy that
is overriding the copy installed in the generated output by Sphinx.)

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>

Former-commit-id: 076e6d3ec85f30eba01e0c5b4094561ade69c710