bt: Don't rely on 'enum bt_buf_type' being 1 byte
De-referencing the pointer from net_buf_user_data(buf) as a pointer to
an enum causes issues on qemu_x86 because the true size is 8-bit, but
the enum is 32-bit on qemu_x86. So we put in a temporary cast to 8-bit
to ensure only 8 bits are read from the pointer.

This fixes a regression from 2fe40201a6a4f0db954d275919a1b21722d1f584 [formerly d3304dc50863ffe5dbc89daa2793c557cf9edd25]
that broke BT on qemu_x86.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>

Former-commit-id: aca1f17dc6810ff708def006c345b8cd5a7cb6ae