boards: Add support for nRF52 Thingy:52 (PCA20020)
A new board configuration is added for Thingy:52, also known as
PCA20020. This includes support for some of the on-board sensors, and
other peripherals such as the on-board GPIO expander.

The board power rail sequencing is also implemented, as the rails need
to be powered up in a correct order and before certain device driver
init calls are made.

Signed-off-by: Aapo Vienamo <aapo.vienamo@iki.fi>

Former-commit-id: 439a63daacec35af8349928baf0bc94a5ca61a12