native test: End tests as soon as they pass fail
To avoid indeterministic behaviour of the tests
which leads to changing coverage reports, let's
finish the tests exectuion as soon as they pass
or fail.
Also let's return 0 to the shell if the test passed,
and 1 if it failed.

(Reverting
f9af42f078c93eca8b9f461f91b6cafddb57a35d and
7225e4a0b73994b31bbd9f76a3222ab8b0e4f40f [formerly 07c9163f67dbbb937cf29680b8243cdc3ed65c07]
)

Signed-off-by: Alberto Escolar Piedras <alpi@oticon.com>

Former-commit-id: 1e46d390684a974b96a54db88368c56d4fa92d51