scripts: nrfjprog.py: Add --softreset switch
The current script relies on a hardware reset pin to release the chip
from reset. Software reset is also possible and even required in case
the reset pin hasn't been wired to the J-Link connector on the board.

This patch adds an additional command line switch to nrfjprog.py that
uses the software reset mechanism of nrfjprog instead of pin reset.

Signed-off-by: Aapo Vienamo <aapo.vienamo@iki.fi>

Former-commit-id: edbaeea22f00146de970d093476d0fdcd3619461