drivers: flash: nrf: Rename nrf5 to nrf
With upcoming ICs that are not in the nRF5x family, rename the flash
driver and all its dependencies from nrf5 to nrf.

Should also fix the issue introduced by 6ce11a48406c047 [formerly f49150cab686ae3] which broke the
assignment of the flash device due to a partial rename.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>

Former-commit-id: 8aa9a37902db4d78925ac3c887ac9409863f0ba8